+++ src/features/validation/calculations.ts
@@ -1,10 +1,12 @@
 import {
   Vector3Array,
+  YawControlPlayer,
   type TrajectoryPlayer,
   type Vector3,
 } from '@skybrush/show-format';
 import { range } from 'lodash-es';
 import { SAMPLES_PER_SECOND } from './constants';
+import {Euler} from 'three';
 
 /**
  * Checks whether two 3D vectors are almost equal within a given epsilon.
@@ -154,4 +156,25 @@ export function sampleVelocityAt(
   const result = new Float32Array(times.length * 3);
   player.getVelocitiesAt(times, result);
   return Vector3Array.from(result);
+}
+
+/**
+ * Samples the yaw of a yaw control player at the given time instants.
+ *
+ * @param player the yaw control player to use for sampling the yaw angles
+ * @param times  the time instants (in seconds) at which to sample
+ */
+export function sampleYawAt(
+  player: YawControlPlayer,
+  times: number[]
+): number[] {
+  const result:number[] = new Array(times.length);
+  // TODO: would be better to define player.getYawsAt() in @skybrush/show-format
+  const temp = new Euler();
+  for (let i = 0; i < times.length; i++) {
+    player.getYawAt(times[i], temp);
+    result[i] = temp.z; // [rad]
+  }
+
+  return result;
 }