+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [main]
+
+### Added
+
+- Added yaw and yaw rate blocks to the validation window. Support for new yaw rate 
+  threshold is also implemented.
+
 ## [2.9.1] - 2026-01-06
 
 ### Fixed